Initial Commit
This header file implements a set of wrapper macros for the
__builtin_load_no_speculate builtin function detailed at
https://www.arm.com/security-update . This builtin function defines a
speculation barrier, which can be used to limit the conditions under which
a value which has been loaded can be used under speculative execution.

The header file provided here allows a migration path to using the builtin
function for users who are unable to immediately upgrade to a compiler which
supports the builtin. Arm recommends using an upgraded compiler where possible
to ensure the most comprehensive support for the mitigation provided by the
builtin function.