Fix handling of pointer-to-const data
Calls to load_no_speculate with a pointer-to-const first argument would fail
to compile. We need to ensure that the type we use within the macros as
an intermediate output for the assembler blocks is suitable to write to.
This patch ensures that for all supported architectures.

The earlier issue with _Bool types has been fixed.