Revert "Fix handling of pointer-to-const data"
This fix would be incorrect. The behaviour in the dispatch switch will change
as a consequence of this change, such that more data than expected will be
returned when the type is bool. As this is a correctness issue which can be
difficult to detect, as opposed to a compile time failure on const data, I
am reverting this immediately.